+++ lib/adaptive_shimmer.dart
@@ -1,3 +1,4 @@
 library adaptive_shimmer;
 
-export './src/adaptive_shimmer.dart';
+export './src/adaptive_shimmer.dart';export 'src/widgets/skeleton_widgets.dart';
+export 'src/models/animation_type.dart';